% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{evaluate}
\alias{evaluate}
\alias{evaluate.r2eng}
\title{Evaluate expression in r2eng object}
\usage{
evaluate(r2eng, ...)

\method{evaluate}{r2eng}(r2eng, envir = parent.frame(), ...)
}
\arguments{
\item{r2eng}{r2eng object to evaluate}

\item{...}{Other arguments to pass}

\item{envir}{environment to evaluate the expression, default to current environment}
}
\value{
Nothing
}
\description{
This function evaluates the expression of the r2eng object.
}
