% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{print.r2eng}
\alias{print.r2eng}
\title{Print r2eng object}
\usage{
\method{print}{r2eng}(r2eng, ...)
}
\arguments{
\item{r2eng}{object to print}
}
\value{
Nothing
}
\description{
Print out the r2eng object.
}
