% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2eng-package.R
\docType{package}
\name{r2eng-package}
\alias{r2eng}
\alias{r2eng-package}
\title{r2eng: Translate R Code To An English Sentence}
\description{
Takes an R expression and converts it to English by matching
    recognised symbols with an opinionated list of English 'translations'.
    Inspired By Amelia McNamara's useR! 2020 conference talk.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matt-dray/r2eng/}
  \item Report bugs at \url{https://github.com/matt-dray/r2eng/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
