% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{r2eng}
\alias{r2eng}
\title{Convert R Code To Spoken English}
\usage{
r2eng(expression, speak = TRUE)
}
\arguments{
\item{expression}{An R expression as a character string.}

\item{speak}{Do you want your system to try and say the English expression?
Requires speakers on your machine.}
}
\value{
A list with three elements: the input R expression, the 'translated'
    English expression and a data frame showing the translation of each
    R element.
}
\description{
Takes an R expression and converts it to English by matching recognised
symbols with an opinionated list of English 'translations'.
}
\examples{
\dontrun{
r2eng("variable <- 1")
}
}
