% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r2eng_dictionary}
\alias{r2eng_dictionary}
\title{An R-To-English Lookup File}
\format{
A data frame with two variables:
\describe{
  \item{r}{an R 'code element', or 'symbol'}
  \item{eng}{the equivalent English term for each R symbol}
}
}
\usage{
r2eng_dictionary
}
\description{
An opinionated data-frame look-up of R symbols to their equivalent English
pronunciation.
}
\keyword{datasets}
