% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{read.r2eng}
\alias{read.r2eng}
\title{Print r2eng Object For Reading}
\usage{
read.r2eng(r2eng, ...)
}
\arguments{
\item{r2eng}{object to print}

\item{...}{Other arguments to pass}
}
\value{
Nothing
}
\description{
Print the r2eng object to see the original and translated expressions.
}
