% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{speak}
\alias{speak}
\alias{speak.r2eng}
\title{Speak out the English expression of r2eng object}
\usage{
speak(r2eng, ...)

\method{speak}{r2eng}(r2eng, ...)
}
\arguments{
\item{r2eng}{An r2eng object to speak.}

\item{...}{Other arguments to pass.}
}
\value{
Nothing.
}
\description{
This function speaks the English expression of the r2eng object. Currently,
    only macOS is supported.
}
