% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate_string}
\alias{translate_string}
\alias{translate}
\title{Translate R Code To Spoken English}
\usage{
translate_string(expression, speak = TRUE, function_call_end = "of ")

translate(expression, speak = TRUE, function_call_end = "of ")
}
\arguments{
\item{expression}{An R expression (bare if \code{translate()}, quoted if
\code{translate_string()}).}

\item{speak}{Logical. Do you want your system to try and say the English
expression? macOS only. Requires speakers on your machine.}

\item{function_call_end}{Character, what should be added after a function
call, e.g. setting this parameter to "of", this function translates
\code{summary(x)} to "summary of x open paren x close paren".}
}
\value{
An r2eng object, which is a list with three elements: the input
    R expression, the 'translated' English expression and a data frame
    showing the translation of each R element.
}
\description{
Takes an R expression and converts it to English by matching recognised
symbols with an opinionated list of English 'translations'.
}
\examples{
\dontrun{
translate(variable <- 1)
}
}
