% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{print.r2eng}
\alias{print.r2eng}
\title{Print r2eng Object For Reading}
\usage{
\method{print}{r2eng}(x, ...)
}
\arguments{
\item{x}{An r2eng object to print.}

\item{...}{Other arguments to pass.}
}
\value{
Nothing.
}
\description{
Print the r2eng object to see the original and translated expressions.
}
